+++ src/app/item-triage/triage-utils.tsx
@@ -355,7 +355,7 @@ function compareBetterStats(
   // this false fallback shouldn't crop up, but just in case, we make no judgement
 }
 
-function isArtifice(item: DimItem) {
+export function isArtifice(item: DimItem) {
   return Boolean(
     item.sockets?.allSockets.some((socket) => socket.plugged?.plugDef.hash === 3727270518)
   );